Refactored /mosaic method to follow DRY
