+++ common/constants/artpieceFilters.js
@@ -1,0 +1,28 @@
+const ArtPieceFilters = {
+  author: {
+    filter: 'not_empty',
+  },
+  title: {
+    filter: 'not_empty',
+  },
+  technique: {
+    filter: 'default',
+  },
+  materials: {
+    filter: 'default',
+  },
+  measurements: {
+    filter: 'default',
+  },
+  year: {
+    filter: 'default',
+  },
+  source: {
+    filter: 'not_empty',
+  },
+  description: {
+    filter: 'default',
+  },
+};
+
+module.exports = ArtPieceFilters;
